Updates publication formats in Thoth
Issue: documentacao-e-tarefas/desenvolvimento_e_infra#873

+++ classes/services/ThothReferenceService.inc.php
@@ -50,6 +50,12 @@ class ThothReferenceService
 
     public function updateReferences($thothClient, $thothReferences, $publication, $thothWorkId)
     {
+        $oldPublication = Services::get('publication')->get($publication->getId());
+
+        if ($publication->getData('citationsRaw') == $oldPublication->getData('citationsRaw')) {
+            return;
+        }
+
         foreach ($thothReferences as $thothReference) {
             $thothClient->deleteReference($thothReference['referenceId']);
         }
